+++ src/condor/event/waves/VanillaWave.java
@@ -8,10 +8,10 @@ import condor.phantom.PhantomType;
 
 public class VanillaWave extends Wave {
 
-  private static TreeMap<PhantomType, Double> map = new TreeMap<>();
+  private static TreeMap<PhantomType, Integer> map = new TreeMap<>();
 
   static {
-    map.put(PhantomType.VANILLA, 3.0);
+    map.put(PhantomType.VANILLA, 3);
   }
 
   public VanillaWave() {
